# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a dynamic Headless CMS API built with Go Fiber and MongoDB.",
        "title": "GoHeadless CMS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticate with username and password, returns a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/collections": {
            "get": {
                "description": "Get metadata for all user-defined collections",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "List all collections",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Collection"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Define a new dynamic collection with fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Create a new collection",
                "parameters": [
                    {
                        "description": "Collection Definition",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Collection"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Collection"
                        }
                    }
                }
            }
        },
        "/collections/{slug}": {
            "get": {
                "description": "Get metadata for a specific collection by its unique slug",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Get collection by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Collection"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a collection definition by its slug",
                "tags": [
                    "collections"
                ],
                "summary": "Delete collection by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/domain.User"
                }
            }
        },
        "domain.AccessControl": {
            "type": "object",
            "properties": {
                "allowed_roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "crud_policy": {
                    "$ref": "#/definitions/domain.CRUDPolicy"
                },
                "is_public": {
                    "type": "boolean"
                }
            }
        },
        "domain.CRUDPolicy": {
            "type": "object",
            "properties": {
                "create": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "delete": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "read": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "update": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "domain.Collection": {
            "type": "object",
            "required": [
                "fields",
                "name",
                "slug"
            ],
            "properties": {
                "access": {
                    "$ref": "#/definitions/domain.AccessControl"
                },
                "description": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Field"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "domain.Condition": {
            "type": "object",
            "properties": {
                "equals": {},
                "field": {
                    "type": "string"
                }
            }
        },
        "domain.ConditionalLogic": {
            "type": "object",
            "properties": {
                "show_if": {
                    "$ref": "#/definitions/domain.Condition"
                }
            }
        },
        "domain.Field": {
            "type": "object",
            "required": [
                "key",
                "type"
            ],
            "properties": {
                "computed_by": {
                    "type": "string"
                },
                "default_value": {},
                "description": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Field"
                    }
                },
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "logic": {
                    "$ref": "#/definitions/domain.ConditionalLogic"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Option"
                    }
                },
                "placeholder": {
                    "type": "string"
                },
                "props": {
                    "description": "Additional dynamic properties for specific components",
                    "type": "object",
                    "additionalProperties": true
                },
                "relation": {
                    "description": "Relational \u0026 Advanced Logic",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.RelationConfig"
                        }
                    ]
                },
                "required": {
                    "type": "boolean"
                },
                "type": {
                    "$ref": "#/definitions/domain.FieldType"
                },
                "unique": {
                    "type": "boolean"
                },
                "validation": {
                    "$ref": "#/definitions/domain.ValidationRules"
                }
            }
        },
        "domain.FieldType": {
            "type": "string",
            "enum": [
                "text",
                "textarea",
                "number",
                "email",
                "password",
                "phone",
                "url",
                "select",
                "radio",
                "checkbox",
                "multiselect",
                "datepicker",
                "timepicker",
                "datetimepicker",
                "file",
                "image",
                "bool",
                "toggle",
                "slider",
                "rating",
                "colorpicker",
                "section",
                "group",
                "tabs",
                "grid",
                "repeater",
                "relation",
                "autocomplete",
                "submit",
                "reset",
                "action"
            ],
            "x-enum-comments": {
                "TypeBool": "Simple boolean"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "Simple boolean",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                ""
            ],
            "x-enum-varnames": [
                "TypeTextInput",
                "TypeTextArea",
                "TypeNumberInput",
                "TypeEmailInput",
                "TypePasswordInput",
                "TypePhoneInput",
                "TypeURLInput",
                "TypeSelect",
                "TypeRadio",
                "TypeCheckbox",
                "TypeMultiSelect",
                "TypeDatePicker",
                "TypeTimePicker",
                "TypeDateTimePicker",
                "TypeFileUpload",
                "TypeImageUpload",
                "TypeBool",
                "TypeToggleField",
                "TypeSliderField",
                "TypeRatingField",
                "TypeColorPickerField",
                "TypeSection",
                "TypeFieldGroup",
                "TypeTabs",
                "TypeGrid",
                "TypeRepeater",
                "TypeRelation",
                "TypeAutocomplete",
                "TypeSubmitButton",
                "TypeResetButton",
                "TypeCustomAction"
            ]
        },
        "domain.Option": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "domain.RelationConfig": {
            "type": "object",
            "properties": {
                "collection": {
                    "type": "string"
                },
                "field": {
                    "description": "UI display field",
                    "type": "string"
                }
            }
        },
        "domain.User": {
            "type": "object",
            "required": [
                "role_id",
                "username"
            ],
            "properties": {
                "active_status": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_initial_admin": {
                    "type": "boolean"
                },
                "role_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "domain.ValidationRules": {
            "type": "object",
            "properties": {
                "max": {
                    "type": "number"
                },
                "max_length": {
                    "type": "integer"
                },
                "min": {
                    "type": "number"
                },
                "min_length": {
                    "type": "integer"
                },
                "regex": {
                    "type": "string"
                }
            }
        },
...
    }
}